refactor: extract class for generate random coins
