+++ src/main/java/vendingmachine/Application.java
@@ -5,6 +5,7 @@ import java.util.Arrays;
 import java.util.List;
 import vendingmachine.model.Coin;
 import vendingmachine.model.Coins;
+import vendingmachine.model.CoinsGenerator;
 import vendingmachine.model.Item;
 import vendingmachine.reader.ExchangeAmountReader;
 import vendingmachine.reader.InputMoneyReader;
@@ -13,31 +14,11 @@ import vendingmachine.reader.ItemListReader;
 public class Application {
 	public static void main(String[] args) {
 		int amount = ExchangeAmountReader.create().read();
-		Coins coins = generateCoins(amount);
+		Coins coins = new CoinsGenerator().generate(amount);
 		printExchangeCoins(coins);
 		List<Item> items = ItemListReader.create().read();
 		int inputMoney = InputMoneyReader.create().read();
 		printInputMoney(inputMoney);
-	}
-
-	private static Coins generateCoins(int amount) {
-		Coins coins = new Coins();
-
-		while (amount != 0) {
-			Coin coin = pickRandomCoin();
-
-			if (amount >= coin.getAmount()) {
-				amount -= coin.getAmount();
-				coins.put(coin);
-			}
-		}
-
-		return coins;
-	}
-
-	private static Coin pickRandomCoin() {
-		int amount = Randoms.pickNumberInList(Arrays.asList(10, 50, 100, 500));
-		return Coin.valueOf(amount);
 	}
 
 	private static void printExchangeCoins(Coins coins) {
