+++ src/main/java/vendingmachine/model/CoinsGenerator.java
@@ -1,0 +1,26 @@
+package vendingmachine.model;
+
+import camp.nextstep.edu.missionutils.Randoms;
+import java.util.Arrays;
+
+public class CoinsGenerator {
+	public Coins generate(int amount) {
+		Coins coins = new Coins();
+
+		while (amount != 0) {
+			Coin coin = pickRandomCoin();
+
+			if (amount >= coin.getAmount()) {
+				amount -= coin.getAmount();
+				coins.put(coin);
+			}
+		}
+
+		return coins;
+	}
+
+	private static Coin pickRandomCoin() {
+		int amount = Randoms.pickNumberInList(Arrays.asList(10, 50, 100, 500));
+		return Coin.valueOf(amount);
+	}
+}
